third_party/blink/web_tests/platform/mac-mac11/external/wpt/performance-timeline/tentative/include-frames-originA-B-expected.txt

This is a testharness.js-based test.
FAIL GetEntries of a parent Frame of origin A and its child frame of origin B assert_equals: Total entries should be 3. expected 3 but got 4
Harness: the test ran to completion.
